refactor: Create src/utils/cache.ts
